Rename sample_data file
