+++ tests/test_routes.py
@@ -12,7 +12,7 @@ from erica.api.v1.endpoints.rente.unlock_code import request_unlock_code, activa
 from erica.pyeric.eric import EricResponse
 from erica.pyeric.pyeric_controller import GetTaxOfficesPyericController
 from erica.request_processing.erica_input.v2.grundsteuer_input import GrundsteuerData
-from tests.sample_data import create_grundsteuer
+from tests.samples.grundsteuer_sample_data import create_grundsteuer
 
 from tests.utils import create_unlock_request, create_unlock_activation, create_est, create_unlock_revocation, \
     missing_cert, missing_pyeric_lib
